requirements.txt
pandas
sqlalchemy
pymssql
plotly
streamlit
PIconnect
python-dotenv
-e .